rust_deoxidizer: move patch into its own file
